feat(sim): begin expanding type system
Move existing TypedBits to expr_value, since it is a value.
Begin constructing a parallel type hierarchy in expr_types.

Need to add some registry for named types (e.g., structs).
Then I can extend the parser to handle better type casts.
Lastly, I can replace TypedBit's type with my new type.
