+++ lib/sim/debug/expr_cache.hpp
@@ -19,11 +19,17 @@ template <typename Item> struct Cache {
     bool operator()(const Item &lhs, const Item &rhs) const { return lhs < rhs; }
   };
   using Set = std::set<std::shared_ptr<Item>, Compare>;
-  template <typename Derived, typename Base = Derived> std::shared_ptr<Base> add_or_return(Derived &&item) {
+  template <typename Derived, typename Base = std::remove_cv_t<std::remove_reference_t<Derived>>>
+  std::shared_ptr<Base> add_or_return(Derived &&item) {
+    // Don't want to make a pointer, and it must be mutable. remove cv and ref.
+    using DerivedType = std::remove_cv_t<std::remove_reference_t<Derived>>;
     QMutexLocker locker(&_mut);
     typename Set::iterator search = _set.find(item);
     if (search == _set.end()) {
-      auto ret = std::make_shared<std::remove_cv_t<Derived>>(std::forward<Derived>(item));
+      std::shared_ptr<DerivedType> ret;
+      // Can't forward lvalue ref?
+      if constexpr (std::is_lvalue_reference_v<Derived>) ret = std::make_shared<DerivedType>(item);
+      else ret = std::make_shared<DerivedType>(std::forward<Derived>(item));
       // Set up dependent tracking on creation. Only called if it exists.
       if constexpr (detail::has_link<Derived>) ret->link();
       _set.insert(ret);
