+++ lib/sim/debug/expr_types.hpp
@@ -1,68 +1,107 @@
 #pragma once
 #include <QString>
 #include <compare>
-#include <concepts>
 #include <cstdint>
-#include <type_traits>
+#include <memory>
+#include "expr_cache.hpp"
 
-namespace pepp::debug {
+namespace pepp::debug::types {
 Q_NAMESPACE
-enum class ExpressionType : uint8_t { i8, u8, i16, u16, i32, u32 };
-Q_ENUM_NS(ExpressionType);
-bool is_unsigned(ExpressionType t);
-// Returns the number of bits in t.
-uint32_t bitness(ExpressionType t);
-ExpressionType common_type(ExpressionType rhs, ExpressionType lhs);
+enum class Primitives : uint8_t { i8, u8, i16, u16, i32, u32 };
+Q_ENUM_NS(Primitives);
+class Type;
+using TypeCache = pepp::debug::Cache<Type>;
 
-struct TypedBits {
-  bool allows_address_of = false;
-  ExpressionType type;
-  uint64_t bits;
-
-  std::strong_ordering operator<=>(const TypedBits &rhs) const;
-  bool operator==(const TypedBits &) const = default;
-
-  friend pepp::debug::TypedBits operator+(const pepp::debug::TypedBits &arg);
-  friend pepp::debug::TypedBits operator-(const pepp::debug::TypedBits &arg);
-  // Operators I do not want to overload because I'm scared I'll break a C++ container
-  friend pepp::debug::TypedBits _deref(const pepp::debug::TypedBits &arg);
-  friend pepp::debug::TypedBits _addressof(const pepp::debug::TypedBits &arg);
-  friend pepp::debug::TypedBits operator!(const pepp::debug::TypedBits &arg);
-  friend pepp::debug::TypedBits operator~(const pepp::debug::TypedBits &arg);
-  friend pepp::debug::TypedBits operator*(const pepp::debug::TypedBits &lhs, const pepp::debug::TypedBits &rhs);
-  friend pepp::debug::TypedBits operator/(const pepp::debug::TypedBits &lhs, const pepp::debug::TypedBits &rhs);
-  friend pepp::debug::TypedBits operator%(const pepp::debug::TypedBits &lhs, const pepp::debug::TypedBits &rhs);
-  friend pepp::debug::TypedBits operator+(const pepp::debug::TypedBits &lhs, const pepp::debug::TypedBits &rhs);
-  friend pepp::debug::TypedBits operator-(const pepp::debug::TypedBits &lhs, const pepp::debug::TypedBits &rhs);
-  friend pepp::debug::TypedBits operator<<(const pepp::debug::TypedBits &lhs, const pepp::debug::TypedBits &rhs);
-  friend pepp::debug::TypedBits operator>>(const pepp::debug::TypedBits &lhs, const pepp::debug::TypedBits &rhs);
-  // Operators I do not want to overload because I'm afraid of interactions with <=> and not being a bool.
-  friend pepp::debug::TypedBits _lt(const pepp::debug::TypedBits &lhs, const pepp::debug::TypedBits &rhs);
-  friend pepp::debug::TypedBits _le(const pepp::debug::TypedBits &lhs, const pepp::debug::TypedBits &rhs);
-  friend pepp::debug::TypedBits _eq(const pepp::debug::TypedBits &lhs, const pepp::debug::TypedBits &rhs);
-  friend pepp::debug::TypedBits _ne(const pepp::debug::TypedBits &lhs, const pepp::debug::TypedBits &rhs);
-  friend pepp::debug::TypedBits _gt(const pepp::debug::TypedBits &lhs, const pepp::debug::TypedBits &rhs);
-  friend pepp::debug::TypedBits _ge(const pepp::debug::TypedBits &lhs, const pepp::debug::TypedBits &rhs);
-  friend pepp::debug::TypedBits operator&(const pepp::debug::TypedBits &lhs, const pepp::debug::TypedBits &rhs);
-  friend pepp::debug::TypedBits operator|(const pepp::debug::TypedBits &lhs, const pepp::debug::TypedBits &rhs);
-  friend pepp::debug::TypedBits operator^(const pepp::debug::TypedBits &lhs, const pepp::debug::TypedBits &rhs);
+class Type : public std::enable_shared_from_this<Type> {
+public:
+  enum class NodeType { Never, Primitive, Pointer, Array, Struct };
+  virtual ~Type() = 0;
+  virtual NodeType node_type() const = 0;
+  virtual bool is_unsigned() const = 0;
+  virtual uint32_t bitness() const = 0;
+  virtual QString to_string() const = 0;
+  virtual std::shared_ptr<Type> common_type(const Type &other, TypeCache &cache) const = 0;
+  virtual std::strong_ordering operator<=>(const Type &rhs) const = 0;
 };
 
-template <std::integral I> TypedBits from_int(I bits) {
-  using T = pepp::debug::ExpressionType;
-  static_assert(sizeof(I) <= 4);
-  T type;
-  auto cast = static_cast<uint64_t>(bits);
-  if constexpr (std::is_same_v<I, int8_t>) type = T::i8;
-  else if constexpr (std::is_same_v<I, uint8_t>) type = T::u8;
-  else if constexpr (std::is_same_v<I, int16_t>) type = T::i16;
-  else if constexpr (std::is_same_v<I, uint16_t>) type = T::u16;
-  else if constexpr (std::is_same_v<I, int32_t>) type = T::i32;
-  else if constexpr (std::is_same_v<I, uint32_t>) type = T::u32;
-  return TypedBits{.allows_address_of = false, .type = type, .bits = cast};
-}
-TypedBits with_bits(const TypedBits &type, uint64_t new_value);
-TypedBits promote(const TypedBits &value, ExpressionType as_type);
-QString to_string(ExpressionType);
+class Never final : public Type {
+public:
+  NodeType node_type() const override;
+  ~Never() override = default;
+  bool is_unsigned() const override;
+  ;
+  uint32_t bitness() const override;
+  ;
+  QString to_string() const override;
+  std::shared_ptr<Type> common_type(const Type &other, TypeCache &cache) const override;
+  std::strong_ordering operator<=>(const Type &rhs) const override;
+  std::strong_ordering operator<=>(const Never &rhs) const;
+};
 
-}; // namespace pepp::debug
+class Primitive final : public Type {
+public:
+  explicit Primitive(Primitives value);
+  Primitive(const Primitive &other) = default;
+  ~Primitive() override = default;
+  NodeType node_type() const override;
+  bool is_unsigned() const override;
+  uint32_t bitness() const override;
+  QString to_string() const override;
+  std::shared_ptr<Type> common_type(const Type &other, TypeCache &cache) const override;
+  std::shared_ptr<Type> common_type(const Primitive &rhs, TypeCache &cache) const;
+  std::strong_ordering operator<=>(const Type &rhs) const override;
+  std::strong_ordering operator<=>(const Primitive &rhs) const;
+  const Primitives value;
+};
+
+class Pointer final : public Type {
+public:
+  explicit Pointer(std::shared_ptr<Type> value);
+  Pointer(const Pointer &other) = default;
+  ~Pointer() override = default;
+  NodeType node_type() const override;
+  bool is_unsigned() const override;
+  uint32_t bitness() const override;
+  QString to_string() const override;
+  std::shared_ptr<Type> common_type(const Type &other, TypeCache &cache) const override;
+  std::strong_ordering operator<=>(const Type &rhs) const override;
+  std::strong_ordering operator<=>(const Pointer &rhs) const;
+
+  const std::shared_ptr<Type> value;
+};
+
+class Array final : public Type {
+public:
+  Array(std::shared_ptr<Type> type, quint16 size);
+  Array(const Array &other) = default;
+  ~Array() override = default;
+  NodeType node_type() const override;
+  bool is_unsigned() const override;
+  uint32_t bitness() const override;
+  QString to_string() const override;
+  std::shared_ptr<Type> common_type(const Type &other, TypeCache &cache) const override;
+  std::shared_ptr<Type> common_type(const Array &other, TypeCache &cache) const;
+  std::strong_ordering operator<=>(const Type &rhs) const override;
+  std::strong_ordering operator<=>(const Array &rhs) const;
+
+  const quint16 size = 0;
+  const std::shared_ptr<Type> value = nullptr;
+};
+
+class Struct final : public Type {
+public:
+  Struct(std::vector<std::shared_ptr<const Type>> members);
+  Struct(const Struct &other) = default;
+  ~Struct() override = default;
+  NodeType node_type() const override;
+  bool is_unsigned() const override { return false; }
+  uint32_t bitness() const override;
+  QString to_string() const override;
+  std::shared_ptr<Type> common_type(const Type &other, TypeCache &cache) const override;
+  std::shared_ptr<Type> common_type(const Struct &other, TypeCache &cache) const;
+  std::strong_ordering operator<=>(const Type &rhs) const override;
+  std::strong_ordering operator<=>(const Struct &rhs) const;
+  const std::vector<std::shared_ptr<const Type>> members;
+};
+
+} // namespace pepp::debug::types
